chore(ui): fix signup page styling (#13086)
* chore(ui): fix signup page styling

* fix the file naming convention

* optimize signup form

(cherry picked from commit 6ef003798bfdbbb7d07a9e9dd3b9dba08dd75745)
