+++ openmetadata-ui/src/main/resources/ui/src/components/AppContainer/AppContainer.test.tsx
@@ -64,7 +64,9 @@ jest.mock('components/AppBar/Appbar', () =>
   jest.fn().mockReturnValue(<p>Appbar</p>)
 );
 
-jest.mock('pages/signup', () => jest.fn().mockReturnValue(<p>SignUpPage</p>));
+jest.mock('pages/SignUp/SignUpPage', () =>
+  jest.fn().mockReturnValue(<p>SignUpPage</p>)
+);
 
 jest.mock('components/router/AuthenticatedAppRouter', () =>
   jest.fn().mockReturnValue(<p>AuthenticatedAppRouter</p>)
