+++ openmetadata-ui/src/main/resources/ui/src/pages/SignUp/account-activation-confirmation.component.tsx
